% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amount_factor.R
\name{amount_factor}
\alias{amount_factor}
\title{Count and identify the factors of a column}
\usage{
amount_factor(x)
}
\arguments{
\item{x}{A dataframe column or vector}
}
\value{
Text showing the amount of factors and which factors. Not useful for piping.
}
\description{
Count and identify the factors of a column
}
\examples{
some_levels <- as.factor(c("A", "A", "B", "C"))
amount_factor(some_levels)
}
