% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_package_name.R
\name{capture_package_name}
\alias{capture_package_name}
\title{Capture package names}
\usage{
capture_package_name(string)
}
\arguments{
\item{string}{string containing "library(package name)"}
}
\value{
string vector
}
\description{
This function takes a typical code chunk containing "library(package name)" as a string as input. It then returns only the package names in this string.
This function is meant to be used with format_version()
}
\examples{
str = "library(tidyverse) # 1.3.1 library(readxl) #1.3.1 library(here) #1.0.1"
capture_package_name(string = str)
}
