% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{common}
\alias{common}
\title{Looks which element of a vector two vectors have in common}
\usage{
common(x, pattern)
}
\arguments{
\item{x}{Vector to search in.}

\item{pattern}{Pattern to find in x. can either be a single element or a vector}
}
\value{
A vector contains the elements that the two vectors have in common.
}
\description{
Looks which element of a vector two vectors have in common
}
\examples{
test1 <- c("A1", "A2", "A3", "A4", "A5")
test2 <- c("A1", "A2", "A3", "B4", "B5")

common(test1, test2)
}
