% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export a dataframe as csv and rds}
\usage{
export(x, path)
}
\arguments{
\item{x}{A dataset to export.}

\item{path}{The file path to export to.}
}
\value{

}
\description{
Export a dataframe as csv and rds
}
