% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_version.R
\name{format_version}
\alias{format_version}
\title{Formats package names}
\usage{
format_version(packages)
}
\arguments{
\item{packages}{string vector}
}
\value{
output in console
}
\description{
This function formats a typical code chunk that loads packages with their current version number. It outputs into the console so that it may be copy pasted into a RMD code chunk.
}
\examples{
pack <- c("dplyr", "utils")
format_version(packages = pack)
}
