% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_inspection.R
\name{general_inspection}
\alias{general_inspection}
\title{#Show basic info about a dataframe}
\usage{
general_inspection(df, x, y)
}
\arguments{
\item{df}{A dataframe}

\item{x}{column to plot on the x-axis}

\item{y}{column to plot on the y-axis}
}
\value{
Summary statistics per column and a basic dot plot.
}
\description{
#Show basic info about a dataframe
}
